+++ src/extend-with-capabilities/All/client/identity/userPanel.js
@@ -12,6 +12,7 @@ class northwindUserPanel extends HTMLElement {
 
         if (!employee) {
 
+            alert("Employee not found; are you in the right tenant?");
             logoff();
 
         } else {
