+++ src/extend-with-capabilities/Monetization/client/identity/userPanel.js
@@ -2,7 +2,8 @@ import {
     getLoggedInEmployee,
     logoff
 } from './identityClient.js';
-
+import { inTeams } from '../modules/teamsHelpers.js';
+import { hasValidLicense } from '../modules/northwindLicensing.js';
 class northwindUserPanel extends HTMLElement {
 
     async connectedCallback() {
@@ -15,6 +16,12 @@ class northwindUserPanel extends HTMLElement {
             logoff();
 
         } else {
+            if (await inTeams()) {
+                const validLicense = await hasValidLicense();
+                if (validLicense.status && validLicense.status.toString().toLowerCase() === "failure") {
+                    window.location.href = `/pages/needLicense.html?error=${validLicense.reason}`;
+                }
+            }
 
             this.innerHTML = `<div class="userPanel">
                 <img src="data:image/bmp;base64,${employee.photo}"></img>
