+++ src/extend-with-capabilities/Monetization/server/validateLicenseService.js
@@ -1,5 +1,9 @@
 import aad from 'azure-ad-jwt';
 import fetch from 'node-fetch';
+import dotenv from 'dotenv';
+dotenv.config();
+
+import { getOboAccessToken } from './identityService.js';
 
 export async function validateLicense(thisAppAccessToken) {
 
@@ -9,7 +13,8 @@ export async function validateLicense(thisAppAccessToken) {
         aad.verify(thisAppAccessToken, { audience: audience }, async (err, result) => {
             if (result) {
                 const licensingAppUrl = `${process.env.SAAS_API}/${process.env.OFFER_ID}`
-                const licensingAppAccessToken = await getOboAccessToken(thisAppAccessToken);
+                const licensingAppAccessToken = await getOboAccessToken(thisAppAccessToken,
+                    process.env.SAAS_SCOPES);
                 if (licensingAppAccessToken === "interaction_required") {
                     reject({ "status":401, "message": "Interaction required"});
                 }
@@ -34,56 +39,5 @@ export async function validateLicense(thisAppAccessToken) {
             }
         });
     });
-
-}
-
-// TODO: Securely cache the results of this function for the lifetime of the resulting token
-async function getOboAccessToken(clientSideToken) {
-
-    const tenantId = process.env.TENANT_ID;
-    const clientId = process.env.CLIENT_ID;
-    const clientSecret = process.env.CLIENT_SECRET;
-    const scopes = process.env.SAAS_SCOPES;
-
-    // Use On Behalf Of flow to exchange the client-side token for an
-    // access token with the needed permissions
-    
-    const INTERACTION_REQUIRED_STATUS_TEXT = "interaction_required";
-    const url = "https://login.microsoftonline.com/" + tenantId + "/oauth2/v2.0/token";
-    const params = {
-        client_id: clientId,
-        client_secret: clientSecret,
-        grant_type: "urn:ietf:params:oauth:grant-type:jwt-bearer",
-        assertion: clientSideToken,
-        requested_token_use: "on_behalf_of",
-        scope: scopes
-    };
-
-    const accessTokenQueryParams = new URLSearchParams(params).toString();
-    try {
-        const oboResponse = await fetch(url, {
-            method: "POST",
-            body: accessTokenQueryParams,
-            headers: {
-                Accept: "application/json",
-                "Content-Type": "application/x-www-form-urlencoded"
-            }
-        });
-
-        const oboData = await oboResponse.json();
-        if (oboResponse.status !== 200) {
-            // We got an error on the OBO request. Check if it is consent required.
-            if (oboData.error.toLowerCase() === 'invalid_grant' ||
-                oboData.error.toLowerCase() === 'interaction_required') {
-                throw (INTERACTION_REQUIRED_STATUS_TEXT);
-            } else {
-                console.log(`Error returned in OBO: ${JSON.stringify(oboData)}`);
-                throw (`Error in OBO exchange ${oboResponse.status}: ${oboResponse.statusText}`);
-            }
-        }
-        return oboData.access_token;
-    } catch (error) {
-        return error;
-    }
 
 }
